refactor: cleanup slideshow utils
